* commit 'e489f97acb1922e5500317374a22c839acd0786c':
  Fix compilation issue with "no-debug" logging
  Cleanup ffdat metric to be able to import rx_throughput.
  Improve combination of DAT speed and L2 throughput values
  Allow ffdat metric to consider rx_throughput field
  Fix token validation
  Improve DAT-Metric by shifting hysteresis to the loss side and consider other layer2 parameters
  Fix fixed integer arithmetics handling
  Track scaling factors of layer2 data elements. Handle different scaling for DLEP conversion.
  Fix router_id for Netjson domain output

+++ include/oonf/libconfig/cfg_tobin.h
@@ -16,7 +16,7 @@ EXPORT int cfg_tobin_strarray(void *reference, size_t bin_size, const struct con
 EXPORT int cfg_tobin_choice(void *reference, size_t bin_size, const struct const_strarray *value,
   const char *(*callback)(size_t idx, const void *ptr), size_t choices_count, const void *ptr);
 EXPORT int cfg_tobin_int(
-  void *reference, size_t bin_size, const struct const_strarray *value, int fractions, size_t int_size);
+  void *reference, size_t bin_size, const struct const_strarray *value, uint16_t fractions, size_t int_size);
 EXPORT int cfg_tobin_netaddr(void *reference, size_t bin_size, const struct const_strarray *value);
 EXPORT int cfg_tobin_acl(void *reference, size_t bin_size, const struct const_strarray *value);
 EXPORT int cfg_tobin_bitmap256(void *reference, size_t bin_size, const struct const_strarray *value);
